Fix ADC read and add debug GPIO outputs.

+++ inc/drivers/at89c51rb2/mcp3008/mcp3008.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 
 // This driver is built to use only the first analog pin for the moment
-uint8_t mcp3008_read_value(void);
+uint16_t mcp3008_read_value(void);
 
 #endif
