+++ inc/hal/hal_adc.h
@@ -2,6 +2,6 @@
 #define HAL_ADC_H__
 #include <stdint.h>
 
-uint8_t adc_read_value(void);
+uint16_t adc_read_value(void);
 
 #endif
