+++ src/hal/at89c51rb2/hal_adc.c
@@ -2,7 +2,7 @@
 
 #include "hal/hal_adc.h"
 
-uint8_t adc_read_value(void)
+uint16_t adc_read_value(void)
 {
     return mcp3008_read_value();
 }
