+++ src/hal/linux/hal_adc.c
@@ -1,6 +1,6 @@
 #include "hal/hal_adc.h"
 
-uint8_t adc_read_value(void)
+uint16_t adc_read_value(void)
 {
     return 23;
 }
